"J.A. Bezemer" <costar@panic.et.tudelft.nl> wrote:
At the moment, I'll do nothing more than this because 1) I have no need
for a serial driver (PCMCIA gets 2MB in just under 4 sec!) and 2) I don't
have time anyway. So I'll leave the programming to any volunteer (and I
understand there may be several). Additions/corrections to my documents
are of course still welcome.

Please feel free to put the docs in the gphoto source tree and/or on the
gphoto website. The mentioned URL is not really a permanent location.


git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@163 67ed7778-7388-44ab-90cf-0a291f65f57c