Updates to make the HPBS variable based on the number of errors that were recieved.  When errors occur, the HPBS is lowered toward the minimum value (512 bytes).  When the error conditions no longer exist, the HPBS is raised toward the maximum value (16384 bytes).
The algorithm could probably use a little work (which is obvious when
you look at the code), but it appears to work well enough for now.
This change was added to attempt to help those people that suffer from
chronic timeout/checksum errors in their packets.  By reducing the HPBS,
we should be reducing the probability that a single packet will be
corrupted.

Remeber, however, that the larger the HPBS, the higher the throughput
because there are fewer "control" bytes passed around.  So, don't think
that you're solving the world's problems by leaving HPBS set at 512
for all eternity.  Some of us actually want a larger value.


git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@206 67ed7778-7388-44ab-90cf-0a291f65f57c