New Canon A50 patch by Werner Almesberger <almesber@lrc.di.epfl.ch>
Fixes a few minor problems and adds downloading of arbitrary files.

The downloading can be reached via Configure/Configure Camera. The window
that appears has the following functions:
 - clicking on any file downloads that file
 - "Clear camera cache" resets all camera information cached in the driver.
   This does _not_ include images already downloaded into gPhoto or such.
   This function is useful if camera and gPhoto have lost synchronization.
 - "Done" simply closed the window

Note that any change in the cached data while the configure window is open
is likely to produce a crash if clicking on a file name. The main purpose
of the download function is to explore the data generated by the camera,
e.g. the .CTG files, so it's not overly user-friendly.

- Werner


git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@248 67ed7778-7388-44ab-90cf-0a291f65f57c