ax203: Fix freespace searching code
Our freespace searching code (for uploading pictures) assumed that
ABFS entries were always written in memory order. It turns out that the
windows software sometimes writes ABFS entries in an other order (probably
when some files were deleted and new files added, these new files then
fill the memory holes of the deleted files, but get added to the end
of the ABFS).

This patch removes the assumption that ABFS entries are always written in
memory order. It also changes ax203_read_v3_3_x_v3_4_x_filecount(), to always
return max file count (other code will then rely on ax203_file_present to
find out which files really are there). This change was made as even though
the windows software does keep a file count stored in ABFS the frame
itself seems to ignore it. So when reading we solely rely on the present / not 
present state just like the frame does. We do still write back the number of
the highest file slot which is used, like windows does.


git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@12928 67ed7778-7388-44ab-90cf-0a291f65f57c