tp6801: move picture renumbering from delete to commit
The picture numbering in the PAT can contain holes from pictures having
been deleted using the frames builtin menu. 

These holes are a problem as if we keep deleting all but the
highest numbered picture and adding new pictures the picture   
number could reach 254 / 255 which have special meaning.

This patch moves the renumbering of pictures from our own delete functon to
the commit function, so that we also fix any holes in the picture numbers
caused by the frames internal delete.


git-svn-id: svn://svn.code.sf.net/p/gphoto/code/trunk@13673 67ed7778-7388-44ab-90cf-0a291f65f57c